fixed some ugly error handling and moved utils
